Add keyboard shortcuts explainer
Resolves #700
Resolves #798
Resolves #985
Resolves #1167
Resolves #1222

This application has had keyboard shortcuts since #603 but they have
been undocumented making them hard to discover and remember.

In this patch we're picking up the work started in #700 and adding
a keyboard shortcut overylay, opened through the Help menu or by
pressing `CmdOrCtrl` + `?`. This dialog shows a list of the
app-specific keyboard shortcuts and what they do.